CONN-10486 One more streaming test
